Fix MISRA C++ Standard Conversions static issues
:Release Notes:
Fix MISRA C++-2008 Rule 4-5-1, Rule 4-5-2, Rule 4-5-3, Rule 4-10-2
static issues

:Detailed Notes:
Rule 4-5-1 is "Expressions with type bool shall not be used as
operands to built-in operators other than the assignment operator=,
the logical operators &&, ||, !, the equality operators == and !=,
the unary & operstor, and the conditional operator."

Rule 4-5-2 is "Expressions with type enum shall not be used as
operands to built-in operators other than the subscript operator
[ ], the assignment operator =, the equality operators == and !=,
the unary & operator, and the relational operators <, <=, >, >="

Rule 4-5-3 is "Expressions with type (plain) char and wchar, shall
not be used as operands to built-in operators other than the
assignment operator =, the equality operators == and !=, and the
unary & operator."

Rule 4-10-2 is"Literal zero (0) shall not be used as the
null-pointer-constant."

:Testing Performed:
QWTC-12386, QWTC-13015, QWTC-11386, QWTC-11385, QWTC-11309

:QA Notes:
N/A

:Issues Addressed:
[WRR-579] Fix 'MISRA C++ Standard Conversions' static issues
	  raised in MCIL

Reviewed-on: http://gpro.lge.com/c/webosose/media-codec-interface/+/427294
Tested-by: Subhasish Mishra <subhasish.mishra@lge.com>

+++ src/impl/v4l2/generic_v4l2_device.cpp
@@ -74,7 +74,9 @@ bool GenericV4L2Device::Open(DeviceType type, uint32_t v4l2_pixfmt) {
     return false;
   }
 
-  device_poll_interrupt_fd_ = eventfd(0, EFD_NONBLOCK | EFD_CLOEXEC);
+  device_poll_interrupt_fd_ =
+      eventfd(0, static_cast<uint32_t>(EFD_NONBLOCK) |
+                     static_cast<uint32_t>(EFD_CLOEXEC));
   if (device_poll_interrupt_fd_ <= 0) {
     MCIL_ERROR_PRINT(": Failed creating a poll interrupt fd");
     return false;
